[CORE-3458] stubbed web service for searching by file name
This implements the API for searching for files defined in the
README.md.  This implementation includes all error relevant error
handing.

For search response, it fakes the search that it always has a single
response with the index set to the lower bound of the results window. 
The resulting file will have a path beginning with /iplant/home and a
name based on the provided name glob with '?'s replaced with 't's and
'*'s removed.